posts/requirements.txt
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt
psycopg2-binary==2.9.10
requests==2.32.3